DatabaseAdministration/Development/ecommerce/product_service/app/schemas.py
from pydantic import BaseModel
from typing import Optional


class BrandBase(BaseModel):
    name: str

class BrandCreate(BrandBase):
    pass

class Brand(BrandBase):
    id: int

    class Config:
        orm_mode = True

class CategoryBase(BaseModel):
    name: str

class CategoryCreate(CategoryBase):
    pass

class Category(CategoryBase):
    id: int

    class Config:
        orm_mode = True

class ProductCreate(BaseModel):
    name: str
    description: str
    price: float
    image_url: str
    category_id: int  # Add category_id to ProductCreate

class ProductUpdate(BaseModel):
    name: Optional[str] = None
    description: Optional[str] = None
    price: Optional[float] = None
    category_id: Optional[int] = None  # Add category_id to ProductUpdate

class Product(ProductCreate):
    id: int

    class Config:
        orm_mode = True
