DatabaseAdministration/Development/ecommerce/product_service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
passlib[bcrypt]
python-jose
python-dotenv
python-multipart